pyproject.toml
[tool.poetry]
name = "america-elections-2024"
version = "0.1.0"
description = ""
authors = ["Henry Post <henryfbp@gmail.com>"]
readme = "README.md"
packages = [{include = "america_elections_2024"}]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.16.0"
openai = "^1.3.8"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
tinydb = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
